Add files via upload
Kuberwhat?? :o always the same reaction when you hear it for the first time.
Yes, it’s Kubernetes! "koo bur NET ease",  which in Greek means pilot of a ship, and going through the definition you can tell why it is named so.
In our attempts of exploring Kubernetes by GCP Tower, we started with the basic step of creating an App and exposing the application using Kubernetes Services.
But what is a Kubernetes service?
A Service in Kubernetes is an abstraction which defines a logical set of Pods and a policy by which to access them. Services enable a loose coupling between dependent Pods. A Service is defined using YAML (preferred) or JSON, like all Kubernetes objects. Although Pods each have a unique IP address, those IPs are not exposed outside the cluster without a Service. Services allow your applications to receive traffic.
Sounds difficult, right? But it is absolutely not! Go through the presentation in which we give the step by step method to implement your own app on GCP! Here we have implemented Kubernetes on Google Container cluster. And yes, it can be done on AWS or Azure as well, and even on bare metal.
Do you have any interesting App, share it with us, we’ll help you to deploy using K8s container orchestrator.